+++ libs/base_core/src/data/list/array/mod.rs
@@ -1,8 +1,11 @@
 // devela_base_core::data::list::array
 //
 #![doc = crate::_DOC_DATA_LIST_ARRAY!()]
+//!
+#![doc = crate::_doc!(extends: array, vec)]
 //
 
+mod ext; // ExtArray, ArrayFmt
 mod from; // ArrayFrom
 mod init; // array_init!
 mod reexports;
@@ -10,6 +13,7 @@ mod reexports;
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
+            ext::*,
             from::*,
             init::*,
             reexports::*,