+++ libs/base_core/src/num/mod.rs
@@ -12,7 +12,7 @@ mod logic; // ConstBool, False, True, const_bool!
 mod ord;
 
 pub mod error; // error types
-pub mod geom;  // (geometry)
+pub mod geom;
 pub mod niche; // NonZero*, NonZero*, NonValue*|NonExtreme*, ne!, nz!
 pub mod quant; // Cycle*, Interval, interval!, Ratio