+++ docs/changelog.md
@@ -200,7 +200,8 @@ Many feature gates are removed in order to make most features make always availa
 
 ### list
 - move to [base]:
-  - type `ConstList`.
+  - types: `ArrayFmt`, `ConstList`.
+  - traits: `ExtArray`.
 - make all `bitfield!` methods consts.
 
 ## lang